// packages/common/content/types/damage-source.ts
export type DamageSourceType =
  | 'all'
  | 'anvil'
  | 'block_explosion'
  | 'campfire'
  | 'charging'
  | 'contact'
  | 'drowning'
  | 'entity_attack'
  | 'entity_explosion'
  | 'fall'
  | 'falling_block'
  | 'fire'
  | 'fire_tick'
  | 'fireworks'
  | 'fly_into_wall'
  | 'freezing'
  | 'lava'
  | 'lightning'
  | 'magic'
  | 'magma'
  | 'none'
  | 'override'
  | 'piston'
  | 'projectile'
  | 'ram_attack'
  | 'self_destruct'
  | 'sonic_boom'
  | 'soul_campfire'
  | 'stalactite'
  | 'stalagmite'
  | 'starve'
  | 'suffocation'
  | 'temperature'
  | 'thorns'
  | 'void'
  | 'wither';